restructured ratings and reviews folder
