+++ client/src/components/Reviews _and_ratings/ratings/Ratings.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import styles from './ratings.module.css';
+
+// Rating summary
+const Ratings = () => (
+  <div className={styles.ratingsContainer}>
+    <div className={styles.ratingsItem}>
+      Rating Summary
+    </div>
+    <div className={styles.ratingsItem}>
+      Breakdown & Recommendations
+    </div>
+    <div className={styles.ratingsItem}>
+      Product Breakdown (Factors)
+    </div>
+  </div>
+);
+
+export default Ratings;
